+++ docs/changelog.rst
@@ -11,6 +11,7 @@ Changelog
 * Tools: fixed opening a project via command line.
 * Tools: fixed importing empty materials from FBX and OBJ models in some cases.
 * Tools: fixed importing roughness and metallic textures from Blender-exported FBX models.
+* Tools: fixed Import button in FBX dialog being erroneously disabled in some cases.
 * Data Compiler: fixed mesh compilers and importers to better detect and avoid empty meshes from reaching the runtime.
 * Lua: fixed ``Quaternion.look()`` returning non-orthonormal rotation matrices.